backend/requirements.txt
# Terra Metrics Dashboard Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
tensorflow==2.20.0
keras==3.11.3
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

# Data Processing
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Satellite Imagery
sentinelhub==3.12.0
rasterio==1.3.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
